WCAG-AA count via CodeSniffer down to zero
Using the browser tool
[CodeSniffer](https://github.com/squizlabs/HTML_CodeSniffer) we were
able to reduce the error count to zero when set to WCAG-AA. Most of
the findings involved missing attributes (alt or id) or anchors
referencing non-existant elements.

To resolve the anchor issue we migrated `react-router` using
`hashHistory` to `browserHistory`. The option to switch back is still
available by setting the environment variable `HASH_ROUTING` to a
value.

In tandem with history change the server(s) have to be able handle
routing to non-existant resources. This has been configured for our
two common scenarios:

 1. via `docker-compose` with the Caddy server
 2. via CloudFoundry `staticfile` buildpack

truetandem/e-QIP-prototype#1300