Fix hit area issue on navigation buttons
So the hit area was fine until the button text wrapped. So scrapped
the use of the CSS table styles and leveraged the CSS flex instead
along with some minor structural changes. The buttons might be one or
two pixels different in width, but nothing visible to the eye.

truetandem/e-QIP-prototype#1388