Address spinner before suggestions are presented
This also refactored several modal usages in to a common component so
it may be reused in different combinations. For example, both the
Introduction and Suggestion component now use the same logic.

The Spinner itself is yet another standalone implementation using the
Modal to allow for usage elsewhere.

truetandem/e-QIP-prototype#1203