Initial payload structure with basic functionality
This removes a lot of old data structures which have shifted in time
and no longer accurately portray the data. These will be completely
removed in later commits.

The Entity interface is the key which combines several other
interfaces to allow for common behavior amongst data structures
without having to explicitly know what the underlying structure is in
every circumstance.

truetandem/e-QIP-prototype#1592