Move the help icon to the field title above
Since the change with the help icon placement being inline with the
field title this instance does not fit the mold anymore because it
does not have a title declared. Moving this up to the original
question.

Also, closed up the gap between the branch question and the location
component.

truetandem/e-QIP-prototype#2050