Allow display of only one email or telephone number
Resolves truetandem/e-QIP-prototype#2096

There is a requirement of one (1) per each type of contact
information. However, we have a requirement to initially display
two (2) items per each.

Previously we set the minimum on the accordion to 2 and then made the
requirement of 1 within the validation logic. This was confusing in
usability testing because once one item was removed the accordion
would recreate it to fulfill the minimum to display requirement. This
happens so quickly it appears as if nothing is happening.