Added addendum branching to history accordions
Resolves truetandem/e-QIP-prototype#1903

Residence, Employment, and Education now have branch questions at the
end of the accordions. The verbiage was taking from the official
guide.

One difference is the Employment section has a follow-up question to
account for section 13c on the guide. If the answer is "yes" then we
zero out the two final branch questions and add a new item to the
accordion.