Removed date field from date_controls
Worked with @floresj on this resolution. Turns out an empty string was
sometimes being sent as the `date` property in JSON. The Go JSON
parsing did not like this and would throw an error. After much
discussion it was concluded this field really wasn't necessary and
instead we can compose the `Date` from the other properties provided.