Update the focus switching when moving to a new section
Resolves truetandem/e-QIP-prototype#2433

The focus switching, albeit a good attempt, did not cover all
exceptions or use cases when the form deviates from the standard
structure (your history sections) or if some content was delayed in
loading. Defaulting to setting the focus to an anchor at the top of
the main content section instead for the time being.