Handle null data from "package" section
Resolves truetandem/e-QIP-prototype#2673

When going to the **Package** section and then navigating to another
section the user would be logged out.

On further investigation the page was exhibiting a full refresh due to
a `null` returning from `sectionData` because there was no data in the
package storage. Accounting for this allows normal functionality with
minimal changes.