Non-focusable SVGs for IE
Resolves truetandem/e-QIP-prototype#2742

In IE (and Edge) the browser would TAB on both the anchor or label as
it was meant to but if an SVG element was a child it would then TAB to
it next instead of the natural order.

IE does not support `tabindex` for SVG elements but does support a
non-standard `focusable` attribute. Also, not all versions of IE
support `focusable` so we also apply a CSS fix to nullify the pointer event.