Create tools to create and verify project checksums
Resolves truetandem/e-QIP-prototype#2637

During the deployment process the artifacts are recompiled in regard
to their designated target. Once the artifacts are collected a
checksum is created of all relevant files which can later be used to
ensure integrity.

To create the checksums manually use `make checksum` and to verify the
results execute `make check`.

This should not be merged in to master until PR #2851