Add SHA256 hash to the archival PDFs
Add the SHA256 hash of the applicant form data to the Author meta-data
field in each archival PDF. This allows the PDF to be linked to the
linked to the corresponding set of application records in eApp. Future
work will move this to a more appropriate custom property in the PDF and
cryptographically sign the PDF via the PDF Signature field. #372