remove DAP
In the [docs for the Digital Analytics
Program](https://github.com/digital-analytics-program/gov-wide-code#appropriate-placement),
they say that "the DAP script tag should not be placed on pages visited
during logged-in sessions". The only pages where the user isn't logged
in are the consent form and the login form, the latter of which we
probably don't want an external script on anyway. Removing DAP for now,
at least until there's a more extensive landing page and other
unauthenticated parts of the site.