Use 2 sizes of the logo to work-around IE11 quirk
IE11 doesn't scale images properly at all and that doesn't mesh
well with this responsive design that sets the logo size based on
window size. There are other work-arounds using Canvas/JavaScript
but that is too much machinery for the moment.

This change uses a 150x150 logo to display on login screen, then
a 80x80 logo for the top banner when navigation menu is displayed.

The NBIS logo will look OK in IE11 until the browser window's width
is too narrow to include the navigation menu, at which point the
small logo will be scaled down from 80x80 and look pretty pixelated
(just in IE).