Update configration to let make specs run on older dockers
Docker versions 18.03.0-ce and later support a Dockerfile being outside
the build context when building; however, versions before that do not.
RHEL7 by default has an older version of docker which doesn't support
this so creating a copy of the file here so that make specs will run.

I created a copy since the file is used else where and I didn't feel it
made sense to change docker-compose.yml to point to a docker file in
specs.