Remove volume from cypress docker image
This is to allow the tests to run in ci since volumes are not setup in
the same way as local machine. This also modifies the make command to
ensure the image is built before tests are run. Also optimize the
dockerfile for cypress to install javascript modules before tests are
copied in