Make Date element serialization more defensive
Similar to https://github.com/18F/e-QIP-prototype/pull/1487. Seems
to be possible to serialize form data in such a way where `month`
property is not set on date. Previous "lazy" version of the code
relied on `month` being set to empty string if not applicable was
indicated for the granted/investigation dates.