Make code more defensive where date can be unknown
Similar to #1490. It seems possible to serialize form data in such
a way that the `month` property is not set on date. This PR makes
the code more defensive in that the month/day/year properties are
not dereferenced if `I don't know` is specified for a date in the UI.