Configure tests to use TEST_DATABASE_NAME to configure which database to configure to.
I also got rid of the database.Configure() method, instead configuring it with a method in the postgres package. This is how all service configuration should work going forward.

This neuters the CheckTable() method, since tests should now run against a real database. I also fixed a bug that was exposed by doing all of this.