Add FOR UPDATE row locking for SaveSection.
The FOR UPDATE clause acquires a lock on all the rows returned by the select query.
The lock makes all UPDATE, DELETE, and SELECT FOR UPDATE queries block until the lock is released.
That lock is released when the transaction it is in commits.
Crucially, it does *not* block naked SELECT clauses, so queries that are just fetching data and aren't
followed up by an update don't get blocked.

Proof is in this gist: https://gist.github.com/macrael/277e5fbdef471c6ae6ccc7c77d3e4f84