branchToAnswer is faulty, excise it from all code.
branchToAnswer will generate a `No` for a branching question in the XML
when the question wasn't even asked. It's unnecesary at this point since
an empty branch will return empty string when called against `branch`,
which will result in empty XML element (e.g., `Answer`) and be removed
in post-template clean-up step.