Convert XMLService to have the templates directory configurable.
This required a surprising amount of change. Because we are using the
template library which takes a map of functions to be passed into the
templates, we have to plumb the configured service all the way through.
To do so we take advantage of golang Method Values, wherein references
to method functions essentially curry the receiver.