Change form versions for consistency with e-QIP
July 2017 is the most recent version of SF-86, which is identical
to the November 2016 version. It is the revision e-QIP references.

The XML XSDs we are using for the 86 and 85 use '2017-07' and '2017-12-draft7'
respectively, so this change creates consistency there as well.

Testing shows that the stored, serialized JSON meta-data for form
version is automatically updated once account.form_version is updated.