Bump birthdates a year for 2020
These date-oriented tests aren't written well and are tied to the
present year from the system clock vs a fixed clock being passed
into the test. This is a minimum viable change to get the tests to
pass again. If it's not properly fixed, the tests will have to be
changed again in 2021, etc.